code/navigation/requirements.txt
actionlib==1.12.1
angles==1.9.12
backports.zoneinfo==0.2.1
base_local_planner==1.16.7
beautifulsoup4==4.12.3
bondpy==1.8.5
cachetools==4.2.4
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.12
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cv_bridge==1.13.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
docopt==0.6.2
dynamic_reconfigure==1.6.5
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
google-api-core==2.8.2
google-auth==2.22.0
google-cloud-speech==2.14.1
googleapis-common-protos==1.56.3
grpcio==1.48.2
grpcio-status==1.48.2
idna==3.7
image_geometry==1.13.1
importlib-resources==5.4.0
interactive-markers==1.11.5
joint_state_publisher==1.12.15
Js2Py==0.74
kdl-parser-py==1.13.3
laser_geometry==1.6.7
message_filters==1.14.13
numpy==1.19.4
packaging==21.3
pipwin==0.5.2
pkg_resources==0.0.0
pocketsphinx==5.0.2
proto-plus==1.23.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pyjsparser==2.7.1
pyparsing==3.1.2
PyPrind==2.11.3
pyqtgraph==0.11.1
pyserial==3.5
pySmartDL==1.3.4
python_qt_binding==0.4.4
pytz-deprecation-shim==0.1.0.post0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.27.1
resource_retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rsa==4.9
rviz==1.13.30
scipy==1.5.4
sensor-msgs==1.12.8
six==1.16.0
smach==2.0.3
smach_ros==2.0.3
smclib==1.8.5
sounddevice==0.4.4
soupsieve==2.3.2.post1
srt==3.5.3
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
topic_tools==1.14.13
tqdm==4.64.1
tzdata==2024.1
tzlocal==4.2
urdfdom-py==0.4.6
urllib3==1.26.18
vosk==0.3.45
webrtcvad==2.0.10
websockets==9.1
xacro==1.13.20
zipp==3.6.0
